commit before pull code

+++ database/migrations/2024_12_27_080704_remove_unique_field_students.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
      public function up()
     {
-        Schema::table('students', function (Blueprint $table) {
-            $table->dropColumn('sso_name');
-        });
+        if (Schema::hasTable('students')) {
+            Schema::table('students', function (Blueprint $table) {
+                $table->dropColumn('sso_name');
+            });
+        }
     }
 
     /**
